day 11 attempt. does not work, runs too long and gets too slow.
I tried doing backtracking with some pruning but this still runs
too slow. It also slows down a lot after running for hours.

My next attempt wil be to modify it to still use the same
algorithm, but not use vectors of strings. Maybe switching
to a floor represented by a 16 bit bitmask will be faster.
Also storing every single path that is tried may be an issue.